refactor(progress): move formatActiveItemCaption to its own file (Phase 6)
ESLint flagged the Phase 3.5f extraction with
`react-refresh/only-export-components` because GlobalProgressBar.tsx
exported both the React component AND the helper function — Fast
Refresh works most reliably when each file has exactly one component
export.

Moves `formatActiveItemCaption` from `components/layout/GlobalProgressBar.tsx`
to `lib/progress-caption.ts`. Same function body, same contract;
the component now imports it. Net effect:
- Fast Refresh works cleanly on GlobalProgressBar edits
- Caption rules can still be unit-tested in isolation
- One fewer eslint warning surfaces in CI

Phase 6 audit summary (all gates green on this branch):
- cargo check: clean
- cargo clippy --lib --no-deps: 0 warnings
- cargo test --lib: 975 passed, 0 failed
- cargo fmt: clean on every file touched by this PR
  (pre-existing fmt drift in commands/{activity_log,dependencies}.rs
  is on origin/main already and out of scope)
- npm run lint: 0 errors, 1 pre-existing warning in updateStore.ts
- npm run type-check: clean
- npm run test: 305 passed
- npm audit --audit-level=high: 0 vulnerabilities (after #708 basic-ftp
  override merged)
